Allow configuration of a hubProtocol
Currently there is the hard-coded protocol 'http' throughout the
node registration process, which means that nodes cannot self-register
to an https hub. This changeset allows for configuration of a
hubProtocol field (and will parse it out of a provided 'hub' parameter
representing the full url).

Additionally this changeset abandons the private struct-like class
'HostPort' in favor of java.net.URL